Makefile: simplify and improve source-installing procedure
In particular, we can create directories on-demand while installing
the files that require them.  The advantage of this strategy is that
allows to test each single bit of the Makefile.  I.e. now one can
run ``make _install_sources`` and that has no implied dependencies
on the ``_install_skel`` rule.

While there, make sure that we bail out of the loop whenever a
command fails.  So, the user is not confused by tons of errors, as
just has happened to me.