Good-enough generator for data to exchange during tests
The initial strategy for the data to exchange was to deploy a very
large file on the server.  The speedtest server would read portions
of it to implement the download test.  The speedtest client would
cache the server response and would use portions of it in order to
implement the upload test.

The came the idea of using PyCrypto's ARC4 to scramble the content
of the seminal BitTorrent and stream tests.

At this point I though it would be nice to use ARC4 to generate
the data to exchange as well.  Just push a set of "AAAA...A" through
and the ARC4 engine will transform it into random data.  Nice, but
then I found it a problem to deploy PyCrypto on the MacOSX (most
likely because I need to learn much more about Python).

So, I googled to search for a cheap and not-so-random way to
generate data.  And I found the illuminating blog post by Jesse
Noller, available at <http://bit.ly/irj8je>.  And here comes
this patch, that implements something like Jesse's idea.

In particular, this patch adds:

1. a random block generator that exploits the core idea from
   Jesse, i.e. to use a collection of words to build the block
   and to rotate the collection each time to add randomicity
   so that the next block is a bit different;

2. a random file class that is very similar to the one implemented
   at neubot/arcfour.py, with the notable that this one employs
   the block generator as the backend.

In addition, this patch modifies Neubot code to use the new
random file class you can find in neubot/blocks.py instead of
the one that lives in neubot/arcfour.py.