Immediately break POLLER.loop() on /api/exit
It doesn't matter if we don't have time to send the response to
the caller.  What matters is that we exit immediately, otherwise
we're going to break neubot uninstall under debian.  There is a
race, indeed, between neubot exiting and debian/control/prerm
trying to remove the '_neubot' user.  If neubot is still running
the neubot user is already logged in.  So we fail to remove the
user and the whole uninstall/reinstall process screws up.