Listener: make the accept() procedure much more robust
Imagine you accept() and then the other end closes the connection
so that when you invoke geetpeername() the transport endpoint is
not connected anymore.  Good morning!  This race condition happens
in the wild and indeed it happened twice in the last 24 hours.

What is particularly annoying is that the code that accepts is
not robust with respect to exceptions triggered by connection_made()
so the effect is that we stop listening.  Ow ow ow!  Very bad since
the error is not with the listening socket but rather with the
newly connected socket.

So, enter in nazi mode and deploy some more robustness into the
accept() procedure.  Now we catch almost everything and pass the
exception to the parent.  An atomic bomb can now explode into
connection_made() without affecting the listening socket.