main.py: more terse --help output
Some bits have been moved to the manual page and also the relative
order of the various forms of the command has been changed to match
the one in the manual page.  In turn, commands in the manual page
are sorted so that the most frequently used one comes first.