net/stream.py: duplicate http/stream.py outgoing functionality
This commit duplicates the functionality of http/stream.py into
net/stream.py, allowing to pass down the chain strings, files and
StringIOs in a transparent way.

The rationale is that I want to employ this functionality in the
BitTorrent code, to decouple the beginning of a PIECE message and
the body, so that I can send and receive very large PIECEs.

I've measured the new code performances.  It seems a bit slower
than the 0.3.7 version of the same code, under MacOSX 10.6.7.  I've
also tweaked the test class, adding some subsequent start_send()
invocations, to ensure that the queue works.