bittorrent/stream.py: allow PIECE block to be a file-like
This commit modifies send_piece() so that it is possible to
pass down the chain both a buffer and a file-like.

We keep separate the two possibilities, i.e. when the block is
a buffer and when the block is a file, because we have measured
that in the buffer case there is a benefit in joining message
preamble and block and passing the joined string down the chain,
while there is a benefit in reading from the file-like when we
have something huge to send.

This patch is very simple thanks to the new send_queue in
net/stream.py, that duplicates the functionality of http/stream.py
outgoing queue.

(Also, note that "!cii" has not padding issues because the '!'
selects the network mode where there is no padding and types have
standard sizes.)

Tested against Neubot 0.3.7 <bittorrent/main.py>, under Ubuntu
GNU/Linux 10.04 without seeing any slowdown.