bittorrent/stream.py: simplify code in the recv path
This commit rewrites the receive loop to be simpler, faster and more
similar to the one in <http/stream.py>.

We have three state variables: self.left, counting the bytes left to read
in the current message; self.buff, used to bufferize data; self.count,
counting the number of bytes in self.buff.

Note that now the handshake is handled as an ordinary message: at the
beginning we set self.left to 68 to read it, and the handshake-handling
code has been moved to _got_message().

This patch has been tested on Ubuntu 10.04 and the new code is faster
in the localhost test.  The new BitTorrent incarnation reaches a speed
of 2.7 Gbit/s on my computer, while the old one peaked at 2.4 Gbit/s.