bittorrent/stream.py: stub for handling jumbo messages
This patch modifies the BitTorrent receiver to be able to deal with
jumbo messages.  We define `jumbo message` a message being bigger than
the maximum size suggested in the BitTorrent protocol specification.

When messages are below the maximum size, we behave as a standard
BitTorrent client, and we bufferize messages.  Instead, when the
message size is bigger than it, we still read the message, but we
stop buffering and we pass upstream the chunks we receive.

This allows to test the BitTorrent protocol using a wider range of
message sizes.  In particular, it allows to do something like our
speedtest test, but using the BitTorrent protocol (the speedtest test
sends very large messages using HTTP).

Note that this patch just adds stubs and actually jumbo messages
are ignored, in the sense that they are passed upstream but they are
not processed.  We will refine support for jumbo messages in
subsequent patches.

This patch has been tested on Ubuntu 10.04 and the few lines of
code we have added apparently do not slow down the receiving loop,
in the localhost test, since we get the same result as before.