bittorrent: Add and use new BitTorrent peer model
This patch adds a new BitTorrent peer model and propagates changes
so that bittorrent now uses the new model.

Note that the new peer model incorporates the functionalities of
the old Upload, Download, and "parent" classes into the same class,
basically because that makes me feel more comfortable with the
source code.

Another relevant change is that the peer model can manage more than
one stream at the same time, so the interface between the stream and
the peer now requires the stream to pass also a reference to self,
for the peer to know what stream received what message.

While there, I have also removed some unused variables from the
stream, and I have renamed the stream in a way that is more consistent
with other streams in Neubot.