bittorrent: cache partial PIECE params in stream.py
When we're reading a jumbo PIECE message and we read more than
the maximum message size, we pass upstream the chunks as we read
them from the socket.

Fine.  But the peer is designed to manage more than one stream
at a time, and we just pass it the chunks without the parameters.
So, we would likely need in the future to cache the PIECE params
for each stream in the peer.

Ugly.  A better solution instead is to cache inside the stream
the current jumbo PIECE parameters and then pass each time such
parameters to the peer.

This is what this patch implements.