bittorrent/peer.py: start measuring after the first PIECE
This patch modifies the measuring algorithm of peer.py to start
measuring after the first piece.  We used to started measuring just
before sending the initial burst of requests.

I believe this is sensible because when we receive the first piece
we can assume that the pipeline is full and is going to be full for
quite some time (of course iff the bdp in the middle is smaller
than the fixed pipeline size we use).