bittorrent/sched.py: ensure we always fill the pipeline
This is quite cheating but... to emulate a continuous transfer we
must be sure that there is a number of messages queued or in flight
between us and our peer.  And this is not always going to happen
if we send just a couple of messages.

OTOH, this is probably going to make the test a bit longer on
slow connections.  Of course I think this is a problem but that
can be solved adapting our parameters to the channel.