net/stream.py: implement close-pending
This patch modifies net/stream.py behavior to defer close
when there is a pending send operation.  This functionality
is already implemented by HTTP and BitTorrent but I want
to move it at the stream level.  The purpose is to put all
the "twisty" stuff in the same place and just implement
protocol decoding and encoding in HTTP and BitTorrent.

I am not sure that this patch is 100% correct, even if it
looks good to me.  I want to push it upstream today when the
release is many days away, so there is time to test it.