net/stream.py: print error if connect doesn't fail immediately
Nonblocking connect() can either complete immediately, returning
success or failure, or you need to poll for writablity on the file
descriptor using select().  The problem is that when the socket
appears to be writable it can be to either because it is actually
connected or due to an error.  To find out we getpeername() and
if that fails we know that it was an error.  Unfortunately we don't
print a LOG.error() message in this case so the user doesn't notice
that connect() failed.

Repair by adding the proper LOG.error() statement in the proper
place.