Update BitTorrent test to avoid simultaneous downloading
This patch updates the BitTorrent test so that first the connector
does a download from the listener, and then the listener does a download
from the connector.

Rationale: simultaneous downloading, as I've measured, produces much
low results.  I guess that's because REQUEST messages are not as light
weight as pure TCP acknowledgments.  I tought that should not have
been a problem, but probably this holds under normal conditions, while
the test is in high load conditions.  (This eventually led me to make
some interesting Facebook state updates, by the way.)