Under debian always use /usr/bin/python
The debian package depends on python >= 2.5 so, no matter how the
system is configured, the package system guarantees that the right
version of python lives at /usr/bin/python.

This patch tries to enforce the above statement adding a custom
neubot executable script for debian-only and modifying the Makefile
to override $BINDIR/neubot with such executable.

Thanks to Seth Johnson for reporting this issue.