neubot/gui: message boxes using various GUIs
While the name seems to imply that we build a mega abstraction,
actually this is just a simple wrapper around the message box.

The purpose is to warn the user with a message box in one of the
following cases:

1. there is an update available from upstream;

2. the user has not changed the default privacy settings.

The general idea is to run a very simple command at system
startup that checks for 1 and 2 and spits out the message box
if needed.