main.py: response.read() before connection.close()
This patch lets main.py behave more gracefully with respect to
the Agent.  In particular now main.py reads the whole response body
before closing the connection.  This avoid an ECONNRESET error on
the Agent.  You may ask "who cares?"  But this one is actually
quite important because this RST pollutes the logs and the user
might wonder why.  And -- that said -- it is always desirable to
behave gracefully, when it does not cost too much.