http/server.py: Do not violate RFC 2616 Sect. 14.30
Sect. 14.30 of RFC 2616 says that Location must return a complete URI
not just a partial URI.  Do that, for increased standard compliancy, even
if this not seems to be the reason why I cannot open the browser from
Neubot in my OpenBSD laptop.

BTW this patch is reasonable as long as we start with the good old IP
version 4, while, with version 6, the address needs to be quoted by square
braces or the browser will fail to parse the URI.  Since sooner or later
we will switch -- whether we like that or not -- deploy an XX comment there
to hint possible future debuggers.