http/message.py: Capitalize HTTP header names
Before sending on the wire, capitalize header names and make sure
that each letter after a dash is capitalized too.  This is how HTTP
headers are sent on the wire by most HTTP clients and servers out
there.  Even if the standard explicitly says that header names aren't
case insensitive, this seems a reasonable thing to do.

By the way, this one too is not the reason why I fail to start the
browser in my OpenBSD latop (the browser seems to complain about mal-
formed HTTP messages).