net/poller.py: Safer timeout algorithm
The problem here is that if a stream is both readble and writeable
and timedout for both the read and the write timeout, then the first
attempt will succeed at removing it while the second will fail, as
the stream has already been removed by the first attempt.

So, rewrite the algorithm for safety and build a set of streams to
be pruned and prune them later.  We use a set so we don't need to care
of adding multiple times the stream to the list.  I.e. using a list
is not going to change the game here (I say this because I initially
committed a version of the code using a list :-P).

Written at the Athens airport.