net: Implement watchdog mechanism
This watchdog mechanism runs together with timeout processing and,
if the watchdog timeout is positive, ensures that the stream is closed
after a given number of seconds.

The mechanism is quite cool because it allows to set up a maximum
life time for transmission tests which is simpler than the current
sched() based approach.

Of course, for this to work, the stream must either be readable or
writable because otherwise the poller does not have an idea that such
stream exists.  While this seems a problem, please note that even in
case of Comet requests the underlying stream is always readable.
Indeed, here in Neubot, we continuously monitor the readability of
a stream (see the code in http/stream.py and bittorrent/stream.py for
exmple).

Written at the Athens airport.