http/server.py: add connection_ready()
This function is invoked when the stream has been created for the
current connection.  This is for simmetry with BitTorrent and with
the HTTP client.  Also, I need this feature to exploit the watch-
dog functionality into the speedtest.  In particular, I have to set
the watchdog on the stream as soon as it is created.

Written at the Athens airport.