http: allow for outgoing body to be a basestring
Yesterday I pushed some code that assumed it was possible to
use a basestring body.  Unfortunately this does not work out of
the box because some points of our HTTP implementation still
assume the body to be a StringIO.  This patch relax the assumption
and now the body can also be a string.  Probably in the next days
some more cases will pop out and we'll deal with them as well.

(Incoming body -- i.e. the one we read from the underlying stream
socket -- continues to be StringIO only: I don't see any advantage
in reading directly into a string, given the complexity of string
concatenation.)