http: allow to rewrite request uri at will
Until today we were not allowed to rewrite request URI because it
was needed to re-build the response line for accounting reasons.
That was a shame because rewriting URIs allows for very beautiful
and twisty things.  Relax the contraint by building the response
line in the Message object itself.  So now we can happily shoot
at our own foots.