negotiate.py: Add new negotiation code
The general idea of the new negotiation code is to split the
negotiation into three layers.  The first layer is the one of
HTTP communication with the client and cares about receiving
and sending HTTP messages only.  The second layer manages the
real negotiation but leaves some details to the third layer.
The third layer manages the per-protocol access list and also
takes care of saving results into the database during the
collect phase.

At the moment this code is unused but it will be picked up
soon by the server-side of BitTorrent.