net/poller.py: of course tasks are AHEAD in time
Bad bad Simone: of course tasks are AHEAD of time, so the
current timeout calculation is broken.  The fix consists of
swapping the two terms of the difference that calculates
the time before the first task expires.

Fixes CPU usage going and staying at 100% on my poor backup
test server and workstation.