speedtest/client.py: anticipate decision of test to perform
This patch anticipates the decision of the test to perform, which
now is taken before checking whether we're allowed to run a test or
not.  Taking the decision earlier allows to log the decided test
even if we're not going to run it.  This helps when debugging the
behavior of Neubot.