bittorrent/client.py: do not collect if test fails
The existing code sends an empty dictionary to the test server
in the collect phase, when the test fails.  This is not the
right thing to do because we should instead tell the other
end that the test failed.

This patch fixes the behavior closing the connection without
sending anything during the collect phase.  Also leave there
a comment to remind that we should tell the other end that
the test failed.

(Of course the other end can figure out by itself that the
test failed but it would be interesting to know what the
connector has seen: this is valuable if someone in the middle
resets both connections.)