state.py: The dumber, the better
We're trying to be too smart here in state.py: to send
selectively what changed only makes the code more complex
because we have to worry about missing an event.  After some
thoughts I think I prefer the data in flight to be more
complex and the code to be simpler.

So revert the optimization I implemented some months ago and
always send the whole dictionary to the client.  Also, do
not send anymore the time of the last change of the variable
(which by the way reduces the size of the JSON).

While there, fetch the definition of T() from utils.py
(where it is declared) and replace int(time.time()) with
utils.timestamp().