state.py: add unit test, remove ad-hoc test code
With the unit test we can be more confident that state.py
keeps behaving well over time.  The ad-hoc test code can be
safely thrown away because it requires an human the read the
results, which is too time consuming.

Please note that we need to make State() more flexible in
order to simplify the unit test.  Namely, now the constructor
needs to know the function to publish() and the function to
fetch the opaque time T().

While there, also use a tuple for state.STATES instead of a
vector, since the former is much faster.

The final thought is that I don't know whether it's a good
idea to fill "pid" and "since" in State(): better to do
this explicitly because the current code cause some oddities
(maked with XXX) in the unit test.