Clear results before test in a robust way
Since it's possible that we miss the negotiate event better
to clear results when entering into the test state, which
should take some seconds.  It's simpler to do this in Neubot
rather than in the Web User Interface (WUI).

Please note that this commit modifies the behavior of speedtest
that now does not report each phase of the test to the WUI.
This reduces the number of Comet updates and is harmful because
the code in js/state.js just cares about current == "test" and
does not look at VALUE in {"test": VALUE}.