negotiate.py: always process request bodies
This patch enhances the negotiator with respect to the
processing of incoming request bodies.

If there's no body the negotiator does not fail and behaves
as the client passed an empty dictionary.

Otherwise the negotiator makes sure that the mimetype is
application/json and that the unserialized object is a
dictionary.