bittorrent/server.py: always use the same PEER_ID
As noted in a previous patch it is quite strange for the
server to generate a new random PEER_ID each time a new client
connects to it.  This patch modifies the code to always use the
same PEER_ID, assuming that PeerNeubot.configure() is called
(this is a reasonable assumption because I doubt bittorrent
would work otherwise).