bittorrent: use same infohash for listener and connector
This commit teaches the listener to wait for the connector
handshake and then reply with an handshake message that
contains the infohash advertised by the connector.  In turn
the connector makes sure that the listener handshake contains
the right infohash.  Now our behavior is more compliant with
BitTorrent specification (BEP 0003).

Also print the infohash when we log the HANDSHAKE message to
aid debugging (when things do not work) and to make it clear
that the two infohashes are equal (when things do work).