bittorrent: unsanely raise watchdog timeout
I don't understand why -- but the upload part of the test takes
too much time and hits too many watchdog timeouts.

This patch raises the timeout insanely to allow to debug the issue
because of course if the watchdog fires the test is over and we cannot
understand much.

Guess tomorrow afternoon I'll have to dig into this issue.