database: By default select ascending
This commit modifies the behavior of the select statement
generator to order by ascending by default and to use descending
on-demand.

Also we need to fix table_bittorrent.py that assumed that the
sorting was descending.

This is needed by the commit that writes logs into the database,
to print log messages in the typical log files order.