bittorrent: improve scheduling and channel adaptation
Adapt the initial burst size to the target bytes instead of
using a fixed number of bytes.  This implementation sends an
initial burst which is 33% of the number of target bytes.
This makes the test shorter for low connections, where the
old fixed-size burst was excessive.

Always adapt the channel using TARGET/elapsed since at the
beginning we're likely to underestimate: if the initial burst
is not big enough we don't even fill the pipeline.  This is
something that won't make the test longer.

Modify the scheduling to be able to request the exact number
of bytes we want to receive, instead of rounding it for excess
on the next message boundary.  This makes the test shorter
because we don't suffer of the rounding error.