bittorrent: quit neubot when the client-side test is finished
Historically, Neubot poller breaks out of the loop when there are
no more file descriptors to monitor for readability or writability.
This behavior is not optimum because it hides some bugs: we want
Neubot to quit when it reaches the final state.  I.e. it does not
matter that the connection with the test server is closed, what
is relevant is that the test code has published the "testdone"
event.

To move into this direction, I have a diff in my pu (proposed
update) git branch that never breaks out of the poller loop.
I run this diff to expose the cases where I need to explicitly
break out of the loop because I've reached the final state.

This diff enhances bittorrent test precisely to do that: now
the client-side test exists upon receiving the "testdone" event.

Eventually I will find all the places when Neubot needs to
break out of the loop, and I can promote the poller diff from
pu to next.  I cannot do that immediately because it's not
so user friendly for the program not to exit when it is
supposed to.