Error slippage: MacOSX getpeername() fails with EINVAL
When the socket becomes writable we use getpeername() to make sure
we're actually connected.  According to Bernstein it should fail with
ENOTCONN when we're not connected, but MacOSX is quite different and
returns EINVAL in this case.  This patch therefore adds a check for
EINVAL too so that we can print meaningful errors to the user.