rendezvous/client.py: Always use fresh configuration
This fixes one of the bugs that have annoyed Alessio and me
mostly: the fact that you need to restart Neubot for the changes
to take effect.  Now it should not be the case anymore because
this diff teaches the rendezvous client to read always from
CONFIG and not from a copy.