agent rendezvous.client: Remove duplicate variables
I've decided to keep the agent ones because those are the
ones that I've already set in various databases and sent out
in various emails.

Please note that I've moved just register_defaults() to
a global scope, so that the variables are defined, and we
have all the positive effects of neubot/config.py but
you don't see them from /api/config because there aren't
the descriptions.  I add the descriptions when you run
the main() of rendezvous.client only.

This will make Alessio Palmero Aprosio very happy.