Make sure CONFIG.register_defaults() scope is always global
So the variables always have a default value even if they
are not visible because their description is not loaded.

More than a month ago, trying to hide unnecessary variables,
I have also removed register_defaults() from the global scope
but this is not a good idea because this prevents variables
to have a default value.  Hiding the description in modules'
main() is enough.