Makefile: quick-dirty fix to adjust permissions
For sure distutils supports a way to enforce permissions on
the installed .py files but, after a quick check, I've not
found it, and so here's a quick-dirty patch to make sure we
ship them with the right permissions.

In particular the problem is that the www/ files are installed
readable by root only and Neubot does not run as root.