Makefile: initial support for uninstalling neubot
From time to time it's better to wipe in particular the installed
modules so that the past (i.e. obsolete modules) does not strike us
in the back.  So deploy an initial command that at least wipes the
modules and the executables, and then in the future we will add
more.