log.py: don't defer commit for too much time
It's risky to defer database.commit() for tens of seconds
because neubot will likely be closed with a commit pending
and this is not clean at all.

Instead keep pending log lines into a queue and write all
of them together into the database when it's time to commit.
Also, when the user wants to read the database, merge the
pending lines in queue with the result too.

While there, deploy some more robustness to be sure that
we're not going to crash if the database has a boo boo.
In this case, don't log the error, at least for now, but
put there a comment to remind.