on_gui_login.py: Loop just once to avoid problems with Windows
Unfortunately breaking out of Tk mainloop() is not that good idea
since we become insensitive to events.  The release is pending and
I don't want to either rewrite the infoboxes or arrange things so
that the infobox is run as an external process.  Today I want to
deploy just the simplest fix/workaround.  In the next cycle we will
deal with this issue so that the on_gui_login program can run in
background permanently.