neubot: Add server.py to start server-side components
For now this command just allows to start all the server-side
components at once and provides knobs to prevent some components
to be loaded.  By default all components are loaded but the
rendezvous because it's not needed on the random test server.

Next step will be to add a very minimal API to this module, so
that a simple Nagios plugin can use it to query Neubot state.