scripts/make_lang_en.js: Autogenerate www/lang/en.js
It is paramount to have www/lang/en.js because people that
wants to translate Neubot needs this file as a starting
point.  We can't indeed ask people to start from the italian
translation of the web pages.

You may ask: Why don't you supply such file in the first
places?  Well the reason is that the files are more readable
with actual text inside.