www/footer.html: always use same value of i18n_resultof
There are multiple instances of i18n_resultof in www/ and not
all the instances have the same value.  This patch modifies the
code so that they always use the same value, i.e. '(n/a)'.

Spotted thanks to the new scripts/make_lang_en.js that expects
all the instances of i18n_resultof in the HTML to have the same
value.