scripts/release: zap code to auto-guess next version number
That piece of code was written when I had the idea that each
subsequent release must be a patch release incrementing the last
number and eventually propagating the change on the left upon
reaching patch==10.

This is no longer the case, with release number jumping due
to reached milestone goals and release candidates (indeed the
support for auto-guessing does not work for -rc releases,
and, even if it worked, it's not clear what the right next
version number should be).

To keep things simple, remove the auto-guess code and now I
have to specify the next version number each time.  While there,
also fix the code that prints usage.