server.py: Expose simple server-side API on port 9775
Unless instructed to do so, the server now listens on port
9775 and implements a server-side API (SAPI).

For now such SAPI just returns 200 Ok if the user requests
either /sapi or /sapi/ and 404 Not Found otherwise.  The next
step will be to write a simple plugin for Nagios that takes
advantage of the SAPI.  Once the infrastructure works, we then
can add more complexity at will.