nagios-plugin-neubot: refuse to run as root
There are two reason why, if running as root, it does not make
sense to drop privileges to the user `_neubot`:

1. that we cannot assume Nagios is running on the same machine
   where Neubot itself is running (it seems more likely that the
   same instance of Nagios monitors a set of servers);

2. that, even if user `_neubot` was available, we should not
   use it because neubot(1) and nagios-plugin-neubot(1) should
   run with different privileges (otherwise if the plugin is
   owned it can kill Neubot or try other nastiness, such as
   reading or writing its memory).

With these two good reasons in mind, and having read that Nagios
usually runs its plugins as the unprivileged user `nagios`, replace
the dropping-privileges code with a simple check that dies if it
finds that the plugin is running as root.  The check is there just
to be sure that everything is working as expected.