server.py: Export server-api using SSL
This commit modifies the server to use SSL for server API
since, probably, the state of the server should be exported
also outside the supposedly safe localhost domain.  Also,
and it goes without saying, this patch modifies API server
to listen to 0.0.0.0 instead of 127.0.0.1.

Of course, all the caveats already expressed in previous
patches and in the source code apply to this patch too.
Basically, if one owns neubot(1) she can steal the private
key without too much effort.