net/CA.py: Generate self-signed certificate under posix
This diff tweaks and revamps net/CA.py so that the root user
can safely use it to generate a self-signed certificate for the
Neubot server under posix.  It also adds the CA command to the
list of commands that neubot(1) main is aware of.

The problem with this diff (and with the current certificate
strategy) is that the output file is owned by the same user that
runs neubot(1).  So, if someone manages to control neubot(1),
she can easily modify and/or steal the certificate.  As mentioned
elsewhere I don't know how to deal with this issue but probably
the answer is hidden somewhere in OpenSSH's sources.