http/server.py: allow to flag some ports as SSL ports
When a port is flagged as SSL, the server will instruct the
network layer to setup a server-side SSL connection.

As you can see this diff hardcodes the name of the certificate
and private key file.  This is annoying but not too much because
for now we expect to use https: only for the Neubot server and
we can just tell people to use that path.

The current solution is flawed in that the file containing the
private key and the certificate file must be readable by the user
that runs neubot server.  I considered using PyOpenSSL to load
the key into SSL_CTX at the beginning, but that is not a silver
bullet because it's still possible for one that owns the process
to read /proc/self/mem and seek for the private key.  I need to
study more in depth this subject before deciding what is the
right(TM) thing to do -- maybe some sort of radical privsep can
save the day.

While there, also add code that prints a warning message when
accept() fails.  It was not relevant before but now it is since
it allows to read eventual SSL errors in the handshake phase.