net/stream.py: Treat zero bytes send() as EOF
Was part of a larger diff, which is not going in for quite
some time, until I get the hang of best practices in private
keys management.  This small diff is not going to hurt, if
you consider that trades a RuntimeError for a quiet close()
of the stream.  At the same time it is good for symmetry with
the read path and because maybe one day we'll use an SSL lib
that can return zero on send() ;-).