New version management strategy
This new version management strategy should be compatible with
the old one (as far as past Neubot versions are concerned), but
has the advantage that it allows to represent versions also in
numeric representation.  E.g.::

        11.44.567-rc1  ==   11.044567001
         canonical             numeric
           repr.                repr.

This allows to save the numeric representation in the database
to write conditionals on the version number with various data
analysis tools.

The numeric representation is similar to the one used by Perl.

Requested by Federico.