neubot: make sure there are no executable files inside
There is no point in keeping executable files below neubot/database
and the recent idea of symlinking to neubot/main/__init__.py is cute
but a wrapper script will do its job as well.

The guarantee that there are not executables below Neubot is sane
because it allows to use find to sanitize permissions and ownership
in a bulk without special cases, when needed.