MacOS: makepkg.py must be run by root.  Period.
The risk of making by mistake a release with incorrect ownership
maybe because I'm tired wins over not requiring root privileges
to ease testing.  Also because makepkg.py is quite stable and does
only need a few tweaks here and there.

While there also update the documentation and clarify that the
result is not zipped but rather tar-gzipped.