updater/unix.py: Updater daemon for Unix
This commits adds neubot/updater/unix.py: the Python3-ready
privilege separated Neubot updater.  The general idea is that
this process will keep running as root.  But the updating is
performed by a short-lived unprivileged child, forked for this
purpose.  Of course, this process will also spawn an ordinary
Neubot agent process.  Which, as before, is not privileged.

When an update is available, the updater process will kill
the unprivileged Neubot agent process.  Then it will install
the update and switch to the new version.  This is done via
the ``BASEDIR/start.sh`` script.