main/__init__.py: ditch creeping featurism
I've never used the start/stop/status arguments and guess
they will be never used.  Moreover, the start command overrides
address and port even if the user does not set them.  This is
a shame because it does not allow the user to override them
via web user interface.  Repair is possible but honestly I
prefer to remove an unused feature than add bloat.