MacOS: Neubot-pkg: Info.plist: omit major and minor
Major and minor in that file are a pain because they are not
updated by scripts/release and manual intervention is needed
whenever major or minor changes.

This commit removes them and we will test the resulting package
under MacOS.  If there are no complains the commit can stay
in tree, otherwise we will revert it.