database/_table_utils.py: more robust INSERT INTO autogen
The problem with our current INSERT INTO is that it assumes
colums follow a given order.  This may be true for tables that
have been created using _table_utils.py, but is likely to be
false for tables created at hand.  The solution is simple:
just tell INSERT_INTO the expected order of the columns.