net/stream.py: undo 262144 recvbuf optimization
Some OSes may be better than others in scaling the receive buffer
and the next release of Neubot will contain code that saves the
OS name into the database.  This allows us to spot the effect of
the operating system on the measurement.

Thanks to Matt Mathis for explaining me why the previous optimization
was not a great idea.