server.py: enable sapi w/o SSL
After some discussion with Federico we agreed that it's not a
disaster to share the server queue length in clear.  Because
a malicious client can estimate it also negotiating and looking
at her position in queue.

Since we don't use SSL we can avoid a lot of work and we can
have the Nagios plugin ready quite soon.

Ok Federico.