www: privacy.js: don't mix 1/0 and true/false in settings.html
This commit forces privacy.js to always use 1/0 instead of true/false
for privacy settings.  The result is that in settings.html privacy
settings are always represented as numbers.  Before they were
represented as numbers until the user changed them, but, from that
point on, they were instead represented as strings.  I think the
new behavior is less confusing.