updater/unix.py: naive chroot for MacOSX
I've found that the ownership of /var/empty is unsafe per the
algorithm used by __chroot().  In particular, the issue is that the
root directory is also writable by the admin group.

Being not a security expert, I studied the algorithm employed by
the OpenSSH guys and implemented something that loosely follows the
same steps.  Basically / /var and /var/empty must be dirs and they
must be owned by and writable only by root.

Unfortunately, now I've not enough knowledge to assess whether
the ownership of / is a real problem or not.  What strikes me is
that MacOSX designers probably does not consider this to be a
real problem.

In the end, I've decided to write a __chroot_naive() that performs
a simpler check: that /var/empty is a directory owned by root and
writable only by root.