updater/unix.py: be more precise with respect to copyright
In the case of __chroot() I had not a clue of what should be
done and so I've more or less translated into Python some code
from the OpenSSH project.  For this reason, this commit adds
at the top of the file the original header found in that file.

With respect to going in background and dropping privileges,
those are stuff I've studied at PoliTO.  So, I initially wrote
the code and then looked at how the same thing is performed
by OpenBSD libc and launchd(8).  After reading their code, I
found that I was not clearing supplementary groups and that
I was not invoking setlogin.  The former can be done quite
easily but the latter, unfortunately, is not possible because
Python does not provide a wrapper for this function in os.